[addrman] Fix new table bucketing during unserialization
An addrman entry can appear in up to 8 new table buckets. We store this
entry->bucket indexing during shutdown so that on restart we can restore
the entries to their correct buckets.

Commit a1a51aac07a6154d1459fd7edce12728f186afd4 broke the
deserialization code so that each entry could only be put in up to one
new bucket. Fix that.