test: fix `assert_debug_log` call-site bugs, add type checks
Two recently added tests (PR #28625 / commit c5da861d76841a0d871cc202702903397457089a
and PR #28634 / commit 1b279b8c45654b498c57705a78144205bea15842)
introduced a bug by wrongly using the `assert_debug_log` helper.
Instead of passing the expected debug string in a list as expected, it
was passed as bare string, which is then interpretered as a list of
characters, very likely leading the debug log assertion pass even if the
intended message is not appearing.

In order to avoid bugs like this in the future, enforce that the
`{un}expected_msgs` parameters are lists.