tracing: cast block_connected duration to nanoseconds
When the tracepoint was introduced in 278344051c4e72e60f647ab7930acf5144a22ebd,
the connect_block duration was passed in microseconds `µs`.
By starting to use steady clock in 8273b4c758134372fe2c0fef0d1c78b5d7df1b99
this changed to nanoseconds `ns`. As the test only checked if the
duration value is `> 0` as a plausibility check, this went unnoticed.

I detected this when setting up monitoring for block validation time
as part of the Great Consensus Cleanup Revival discussion.

This change casts the duration explicitly to nanoseconds (as it has been
nanoseconds for the last three releases; switching back now would 'break'
the broken API again; there don't seem to be many users affected), updates
the documentation and adds a check for an upper bound to the tracepoint
interface tests. The upper bound is quite lax as mining the block takes
much longer than connecting the empty test block. It's however able to
detect incorrect duration units passed.